[#57] fix : ArtWork Play  index.js에서 markup 완전히 분리
