+++ front/src/App/pages/Home/container/ArtWorkContainer.scss
@@ -1,0 +1,15 @@
+.ArtWorkPlayContainer {
+  &-category {
+    display: flex;
+    padding: 70px 65px 20px 65px;
+    font-size: 23px;
+  }
+  &-category-title {
+    margin-left: 20px;
+  }
+  &-songWrapper {
+    padding: 0px 75px;
+    display: flex;
+    justify-content: space-between;
+  }
+}
