+++ front/src/App/pages/Home/index.js
@@ -1,4 +1,4 @@
-import React, { Component } from "react"
+import React, { PureComponent } from "react"
 import { connect } from "react-redux"
 import { changePlayList } from "../../../redux/playlist/actions"
 import Navigation from "./components/Navigation/index"
@@ -8,75 +8,25 @@ import { loadSongsInfo } from "src/redux/music/actions"
 import SONG_URL_LIST from "../../constants/test/SongUrlConstants"
 import SONG_URL_LIST1 from "../../constants/test/SongUrlConstants1"
 import { resolveUrl } from "../../constants/ApiConstants"
-//import ArtworkPlay from "./components/ArtworkPlay"
 import ArtWorkContainer from "../Home/container/ArtWrokContainer"
 import classnames from "classnames/bind"
 import css from "./index.scss"
 
-import selectIcon from "../../../assets/icons/icon2.png"
-
 const cx = classnames.bind(css)
 const moduleName = "Home"
 
-const activePalyList = {
-  cursor: "pointer",
-  background: `url(${selectIcon}) no-repeat -49px -152px`,
-  width: "35px",
-  height: "35px"
-}
-
-const inactivePalyList = {
-  cursor: "pointer",
-  background: `url(${selectIcon}) no-repeat -111px -152px`,
-  width: "35px",
-  height: "35px"
-}
-
-class Home extends Component {
+class Home extends PureComponent {
   constructor(props) {
     super(props)
     this.state = {
-      sebaSongInfo: [],
-      knowSongInfo: [],
       sebaChoiceActive: false,
       knowListActive: false
     }
   }
   componentDidMount() {
-    this.props.loadSongsInfo(SONG_URL_LIST)
-    // this.props.loadSongsInfo(SONG_URL_LIST1)
+    this.props.loadSongsInfo(SONG_URL_LIST1)
   }
 
-  requestKnowSongInfo = () => {
-    SONG_URL_LIST1.map(url => {
-      return axios.get(resolveUrl(url)).then(response => {
-        this.setState(prevState => {
-          return {
-            ...prevState,
-            knowSongInfo: [...prevState.knowSongInfo, response.data]
-          }
-        })
-      })
-    })
-  }
-  /*
-  requestSongInfo = (songUrl, infoArray) => {
-    console.log("request song info ", infoArray)
-    songUrl.map(url => {
-      return axios.get(resolveUrl(url)).then(response => {
-        //2d array state ?
-        this.setState(prevState => {
-          console.log("request song info ", infoArray)
-          console.log("request song info prevState ", prevState)
-          return {
-            ...prevState,
-            infoArray: [...prevState.infoArray, response.data]
-          }
-        })
-      })
-    })
-  }
-  */
   //plya list : kind of music list
   dispatchPlayList = playlist => {
     //this.props.changePlayList()
@@ -131,87 +81,8 @@ class Home extends Component {
       <div className={cx(`${moduleName}`)}>
         <Navigation />
         <div>
-          <div className={cx(`${moduleName}-category`)}>
-            <div
-              style={
-                this.state.sebaChoiceActive ? activePalyList : inactivePalyList
-              }
-              onClick={() => {
-                this.dispatchPlayList("SEBA_CHOICE")
-              }}
-            />
-            <div className={cx(`${moduleName}-category-title`)}>
-              SEBA's Choice
-            </div>
-          </div>
-          <div className={cx(`${moduleName}-songWrapper`)}>
-            <ArtWorkContainer />
-          </div>
-
-          <div className={cx(`${moduleName}-category`)}>
-            <div
-              style={
-                this.state.knowListActive ? activePalyList : inactivePalyList
-              }
-              onClick={() => {
-                this.dispatchPlayList("YOU_KNOW")
-              }}
-            />
-
-            <div className={cx(`${moduleName}-category-title`)}>
-              Artists you should know
-            </div>
-          </div>
-          <div className={cx(`${moduleName}-songWrapper`)} />
+          <ArtWorkContainer category="SEBA'S CHOICE" />
         </div>
-        {/*
-        {this.state.sebaSongInfo.length !== 6 ||
-        this.state.knowSongInfo.length !== 6 ? (
-          <Loading />
-        ) : (
-          <div>
-            <div className={cx(`${moduleName}-category`)}>
-              <div
-                style={
-                  this.state.sebaChoiceActive
-                    ? activePalyList
-                    : inactivePalyList
-                }
-                onClick={() => {
-                  this.dispatchPlayList("SEBA_CHOICE")
-                }}
-              />
-              <div className={cx(`${moduleName}-category-title`)}>
-                SEBA's Choice
-              </div>
-            </div>
-            <div className={cx(`${moduleName}-songWrapper`)}>
-              {this.state.sebaSongInfo
-                ? this.rederDiscover(this.state.sebaSongInfo)
-                : "Loading"}
-            </div>
-
-            <div className={cx(`${moduleName}-category`)}>
-              <div
-                style={
-                  this.state.knowListActive ? activePalyList : inactivePalyList
-                }
-                onClick={() => {
-                  this.dispatchPlayList("YOU_KNOW")
-                }}
-              />
-              <div className={cx(`${moduleName}-category-title`)}>
-                Artists you should know
-              </div>
-            </div>
-            <div className={cx(`${moduleName}-songWrapper`)}>
-              {this.state.knowSongInfo
-                ? this.rederDiscover(this.state.knowSongInfo)
-                : "Loading"}
-            </div>
-          </div>
-        )}
-              */}
       </div>
     )
   }
